NEXUS-6185: Applied fixes
Changes:
* reformat of code
* removal of irrelevant comments and javadoc tags
* removal of added dependency in POM, is coming as transitive
* UT made one class, dummy repo made inner of UT

No code change happened to original changes (not counting
the move of dummy repo into UT class)
